add flag file
